feat: Implement Trigger and WASDTrigger classes for key binding functionality
- Added Trigger class to handle key and modifier combinations with validation.
- Introduced WASDTrigger class for specific WASD key mappings to directions.
- Created utility classes for template generation and path management.
- Refactored existing code to utilize new Trigger and template structures.
- Removed deprecated files and consolidated functionality into new modules.
- Updated imports across the codebase to reflect new structure.
- Added error handling for invalid key and modifier inputs.
