+++ CityOfBinds/__init__.py
@@ -1,0 +1,3 @@
+from .src import *
+
+__version__ = "0.1.0"
