+++ CityOfBinds/src/BindFile/constants.py
@@ -1,0 +1,3 @@
+class BindFileConstants:
+    MAX_LINES = 242 # TODO: 25-11-24: need to verify this value
+    EXTENSION = ".txt"
