+++ CityOfBinds/src/RotatingBinds/randomwalk.py
@@ -1,0 +1,18 @@
+import random
+from CityOfBinds.src.Binds.bind import Bind
+from CityOfBinds.src.SlashCommands.slashcommand import Commands
+from CityOfBinds.src.BindFile.bindfile import BindFile, BindFileNode, BindFileGraph
+from CityOfBinds.src.BindFileGraphPublisher.publisher import BFGPublisher
+
+class RandomWalk(BFGPublisher):
+
+    def __init__(self, wasd_template: list = None, jump_template: list = None, *args, **kwargs):
+        self.wasd_template = wasd_template
+        self.jump_template = jump_template
+
+    def _build_binds(self, trigger, template):
+        # Implementation to build binds based on the trigger and template
+        pass
+
+    def _create_nodes(self) -> list[BindFileNode]:
+        pass
