+++ CityOfBinds/utils/Templates/pool.py
@@ -1,0 +1,72 @@
+import random
+from enum import Enum
+
+class SelectionType(Enum):
+    SEQUENTIAL = 0
+    RANDOM = 1
+    RANDOM_ORDER = 2
+
+class Pool:
+    DEFAULT_SELECT_BEHAVIOR = SelectionType.SEQUENTIAL
+    DEFAULT_IS_FINITE = False
+    DEFAULT_RANDOM_SEED = 0xDEADBEEF
+
+    def __init__(
+        self, 
+        name: str, 
+        items: list, 
+        select_behavior: SelectionType = DEFAULT_SELECT_BEHAVIOR, 
+        is_finite: bool = DEFAULT_IS_FINITE,
+        random_seed: int = DEFAULT_RANDOM_SEED,
+    ):
+        self.name: str = name
+        self.items: list = items # TODO: verify list input, decide if it should be mutable (2025/11/27) 
+        self.is_finite: bool = is_finite
+        self._select_type: SelectionType = select_behavior
+        self.random_seed: int = random_seed
+
+        self._access_count = 0
+        self._random_gen = random.Random(self.random_seed)
+        self._random_items = self._new_random_order(self.items)
+
+    @property
+    def access_count(self) -> int:
+        return self._access_count
+
+    def pop(self):
+        item = self.peek()
+        if item is not None:
+            self._pop_update()
+        return item
+    
+    def peek(self):
+        if self._select_type == SelectionType.RANDOM:
+            self._random_gen.seed(self._access_count + self.random_seed)
+            return self._random_gen.choice(self.items)
+
+        if self.is_finite and self._access_count == len(self.items):
+            return None
+
+        if self._select_type == SelectionType.SEQUENTIAL:
+            return self.items[self._access_count % len(self.items)]
+        
+        if self._select_type == SelectionType.RANDOM_ORDER:
+            """ creates a new random list whenever random list is exhausted """
+            return self._random_items[self._access_count % len(self.items)]
+
+    def set_select_behavior(self, behavior: SelectionType): # TODO: do I really want this value to be updated? Does it cause weird behavior as is? (2025/11/27) 
+        self._select_type = behavior
+
+    def _pop_update(self):
+        self._access_count += 1
+        if self._select_type == SelectionType.RANDOM_ORDER:
+            if self._access_count % len(self.items) == 0:
+                self._random_items = self._new_random_order(self.items)
+
+    def _new_random_order(self, items: list) -> list:
+        random_items = items.copy()
+        self._random_gen.shuffle(random_items)
+        return random_items
+
+    def __str__(self):
+        return f"{StringTemplate.ENCAPSULATION_LEFT}{self.name}{StringTemplate.ENCAPSULATION_RIGHT}"
