--- src/CityOfBinds/__init__.py
@@ -1,26 +1,0 @@
-from .trigger import Trigger, WASDTrigger, TriggerMixin
-from .slashcommand import SlashCommand
-from .commandgroup import CommandGroup
-from .power import Power
-from .binds import Bind, WASDBind
-from .comments import Comment, CommentBanner
-from .bindfile import BindFile, BindFileLinker
-from .RotatingBinds.rotatingbind import RotatingBind
-from .pathgenerator import BaseConverter, PathGenerator
-
-__all__ = ["Trigger",
-           "WASDTrigger",
-           "TriggerMixin",
-           "SlashCommand",
-           "CommandGroup",
-           "Power",
-           "Bind",
-           "WASDBind",
-           "Comment",
-           "CommentBanner",
-           "BindFile",
-           "BindFileLinker",
-           "RotatingBind",
-           "BaseConverter",
-           "PathGenerator"
-           ]
